Move pos/pot data structures into corresponding modules
