+++ crates/pallet-subspace/src/tests.rs
@@ -43,8 +43,9 @@ use std::num::NonZeroU32;
 use std::sync::{Arc, Mutex};
 use subspace_core_primitives::crypto::Scalar;
 use subspace_core_primitives::pieces::PieceOffset;
+use subspace_core_primitives::pot::PotOutput;
 use subspace_core_primitives::segments::SegmentIndex;
-use subspace_core_primitives::{PotOutput, PublicKey, RewardSignature, SolutionRange};
+use subspace_core_primitives::{PublicKey, RewardSignature, SolutionRange};
 use subspace_runtime_primitives::{FindBlockRewardAddress, FindVotingRewardAddresses};
 
 #[test]
