+++ crates/sc-proof-of-time/src/source.rs
@@ -24,7 +24,8 @@ use std::collections::HashSet;
 use std::marker::PhantomData;
 use std::sync::Arc;
 use std::thread;
-use subspace_core_primitives::{PotCheckpoints, PublicKey};
+use subspace_core_primitives::pot::PotCheckpoints;
+use subspace_core_primitives::PublicKey;
 use thread_priority::{set_current_thread_priority, ThreadPriority};
 use tokio::sync::broadcast;
 use tracing::{debug, error, trace, warn};
