+++ crates/sc-proof-of-time/src/source/timekeeper.rs
@@ -6,7 +6,7 @@ use futures::SinkExt;
 use sp_consensus_slots::Slot;
 use std::num::NonZeroU32;
 use std::sync::Arc;
-use subspace_core_primitives::{PotCheckpoints, PotSeed};
+use subspace_core_primitives::pot::{PotCheckpoints, PotSeed};
 use subspace_proof_of_time::PotError;
 use tracing::{debug, trace};
