+++ crates/sp-consensus-subspace/src/lib.rs
@@ -42,12 +42,13 @@ use sp_runtime_interface::{pass_by, runtime_interface};
 use sp_std::num::NonZeroU32;
 #[cfg(feature = "std")]
 use subspace_core_primitives::crypto::kzg::Kzg;
+use subspace_core_primitives::pot::{PotCheckpoints, PotOutput, PotSeed};
 use subspace_core_primitives::segments::{
     HistorySize, SegmentCommitment, SegmentHeader, SegmentIndex,
 };
 use subspace_core_primitives::{
-    Blake3Hash, BlockHash, BlockNumber, PotCheckpoints, PotOutput, PotSeed, PublicKey,
-    RewardSignature, SlotNumber, Solution, SolutionRange,
+    Blake3Hash, BlockHash, BlockNumber, PublicKey, RewardSignature, SlotNumber, Solution,
+    SolutionRange,
 };
 #[cfg(feature = "std")]
 use subspace_proof_of_space::chia::ChiaTable;
