+++ crates/sp-domains/src/lib.rs
@@ -68,7 +68,8 @@ use sp_weights::Weight;
 #[cfg(feature = "std")]
 use std::collections::BTreeSet;
 use subspace_core_primitives::crypto::blake3_hash;
-use subspace_core_primitives::{bidirectional_distance, Blake3Hash, PotOutput, Randomness, U256};
+use subspace_core_primitives::pot::PotOutput;
+use subspace_core_primitives::{bidirectional_distance, Blake3Hash, Randomness, U256};
 use subspace_runtime_primitives::{Balance, Moment};
 
 /// Key type for Operator.
