+++ crates/subspace-farmer-components/benches/proving.rs
@@ -14,9 +14,10 @@ use subspace_archiving::archiver::Archiver;
 use subspace_core_primitives::crypto::kzg;
 use subspace_core_primitives::crypto::kzg::Kzg;
 use subspace_core_primitives::pieces::Record;
+use subspace_core_primitives::pos::PosSeed;
 use subspace_core_primitives::sectors::SectorId;
 use subspace_core_primitives::segments::{HistorySize, RecordedHistorySegment};
-use subspace_core_primitives::{Blake3Hash, PosSeed, PublicKey, SolutionRange};
+use subspace_core_primitives::{Blake3Hash, PublicKey, SolutionRange};
 use subspace_erasure_coding::ErasureCoding;
 use subspace_farmer_components::auditing::audit_plot_sync;
 use subspace_farmer_components::file_ext::{FileExt, OpenOptionsExt};
