+++ crates/subspace-proof-of-space/src/shim.rs
@@ -4,7 +4,8 @@
 use crate::{PosTableType, Table, TableGenerator};
 use core::iter;
 use subspace_core_primitives::crypto::blake3_hash;
-use subspace_core_primitives::{PosProof, PosSeed, U256};
+use subspace_core_primitives::pos::{PosProof, PosSeed};
+use subspace_core_primitives::U256;
 
 /// Subspace proof of space table generator.
 ///
