+++ crates/subspace-proof-of-time/src/aes/x86_64.rs
@@ -1,6 +1,6 @@
 use core::arch::x86_64::*;
 use core::mem;
-use subspace_core_primitives::PotCheckpoints;
+use subspace_core_primitives::pot::PotCheckpoints;
 
 /// Create PoT proof with checkpoints
 #[target_feature(enable = "aes")]
