+++ crates/subspace-service/src/lib.rs
@@ -127,7 +127,8 @@ use std::sync::Arc;
 use std::time::Duration;
 use subspace_core_primitives::crypto::kzg::{embedded_kzg_settings, Kzg};
 use subspace_core_primitives::pieces::Record;
-use subspace_core_primitives::{BlockNumber, PotSeed, PublicKey, REWARD_SIGNING_CONTEXT};
+use subspace_core_primitives::pot::PotSeed;
+use subspace_core_primitives::{BlockNumber, PublicKey, REWARD_SIGNING_CONTEXT};
 use subspace_erasure_coding::ErasureCoding;
 use subspace_networking::libp2p::multiaddr::Protocol;
 use subspace_networking::utils::piece_provider::PieceProvider;
