+++ shared/subspace-proof-of-space-gpu/src/cuda.rs
@@ -9,7 +9,7 @@ use rust_kzg_blst::types::fr::FsFr;
 use std::ops::DerefMut;
 use subspace_core_primitives::crypto::Scalar;
 use subspace_core_primitives::pieces::Record;
-use subspace_core_primitives::{PosProof, PosSeed};
+use subspace_core_primitives::pos::{PosProof, PosSeed};
 
 extern "C" {
     /// # Returns
