+++ test/subspace-test-service/src/lib.rs
@@ -83,7 +83,8 @@ use std::marker::PhantomData;
 use std::pin::Pin;
 use std::sync::Arc;
 use std::time;
-use subspace_core_primitives::{BlockNumber, PotOutput, PublicKey, Solution};
+use subspace_core_primitives::pot::PotOutput;
+use subspace_core_primitives::{BlockNumber, PublicKey, Solution};
 use subspace_runtime_primitives::opaque::Block;
 use subspace_runtime_primitives::{AccountId, Balance, Hash, Signature};
 use subspace_service::transaction_pool::FullPool;
